requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
Pillow==10.1.0
python-dotenv==1.0.0
django-filter==23.5
django-guardian==2.4.0
django-storages==1.14.2
celery==5.3.6
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1